requirements.txt
asgiref==3.11.0
django==6.0
django-allauth==65.13.1
django-htmx==1.27.0
djangorestframework==3.16.1
sqlparse==0.5.4
